Move URL variable declaration
Move the url variable declaration into the Launch Handler. This will prevent errors if the sample is mostly used as-is and videoUrl is set to a presigned URL to a developer supplied video hosted in an Alexa-hosted skill.  The pre-signed URL expires after a minute, so if the Lambda container's lifetime is longer than one minute, then the skill response will vend an expired URL. This would be fine (simply check the logs, fix the problem), however the error message is only on the client side (i.e. Echo Show) with nothing in CloudWatch logs, making it hard to debug, especially for someone new to skill development.